Reverted midi parser promotion.
