+++ demos/midi/sendRaw/index.html
@@ -53,7 +53,6 @@
         <script src="../../../src/web/input-device-manager.js"></script>
         <script src="../../../src/web/midi.js"></script>
         <script src="../../../src/midi/controller.js"></script>
-        <script src="../../../src/midi/rawMIDIParser.js"></script>
         <script src="../../../src/web/native-node-manager.js"></script>
         <script src="../../../src/web/output-manager.js"></script>
         <script src="../../../src/parser.js"></script>
