+++ index.js
@@ -32,7 +32,6 @@ require("./src/flocking-audiofile-compatibility.js");
 require("./src/scheduler.js");
 require("./src/web/midi.js");
 require("./src/midi/controller.js");
-require("./src/midi/rawMIDIParser.js");
 
 // Unit generators
 require("./src/ugens/core.js");
