+++ tests/unit/html/midi-tests.html
@@ -26,7 +26,6 @@
         <script src="../../../src/web/buffer-writer.js"></script>
         <script src="../../../src/web/input-device-manager.js"></script>
         <script src="../../../src/web/midi.js"></script>
-        <script src="../../../src/midi/rawMIDIParser.js"></script>
         <script src="../../../src/web/native-node-manager.js"></script>
         <script src="../../../src/web/output-manager.js"></script>
         <script src="../../../src/buffers.js"></script>
